requirements.txt
# pip install -r requirements.txt

# base -------------------------------
matplotlib>=3.2.2
numpy>=1.20.3
opencv-python>=4.4.0
pillow
pyyaml>=5.3.1
scipy>=1.5.2
torch>=1.8.1
torchvision>=0.9.1
tqdm>=4.51.0

# logging ----------------------------
tensorboard>=2.5.0
wandb>=0.10.31

# plotting ---------------------------
seaborn>=0.11.1
pandas

# export -----------------------------
# onnx>=1.9.0

# extras -----------------------------
pycocotools>=2.0  # COCO mAP
thop  # FLOPS computation